backend/api/requirements.txt
fastapi==0.115.0
mangum==0.17.0
py-getstream==1.0.7
python-dotenv==1.0.0
uvicorn==0.30.0